refactor: separate custom style logics in context menu to hook
