+++ src/inputs/context-menu/story-helper.ts
@@ -112,115 +112,6 @@ export const getContextMenuArgTypes = (): ArgTypes => ({
             type: 'boolean',
         },
     },
-    autoHeight: {
-        name: 'autoHeight',
-        type: { name: 'boolean' },
-        description: 'If you don\'t want the context menu to be long enough to go outside the window area, set this value to true.',
-        defaultValue: false,
-        table: {
-            type: {
-                summary: 'boolean',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: false,
-            },
-        },
-        control: {
-            type: 'boolean',
-        },
-    },
-    useCustomStyle: {
-        name: 'useCustomStyle',
-        type: { name: 'boolean' },
-        description: 'Only when this value is true, style can be customized using width, height, offsetTop, and position props.',
-        defaultValue: false,
-        table: {
-            type: {
-                summary: 'boolean',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: false,
-            },
-        },
-        control: {
-            type: 'boolean',
-        },
-    },
-    position: {
-        name: 'position',
-        type: { name: 'string' },
-        description: `Menu position. ${Object.values(CONTEXT_MENU_POSITION)} are available.`,
-        defaultValue: undefined,
-        table: {
-            type: {
-                summary: 'string',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: 'undefined',
-            },
-        },
-        control: {
-            type: 'select',
-            options: [undefined, ...Object.values(CONTEXT_MENU_POSITION)],
-        },
-    },
-    offsetTop: {
-        name: 'offsetTop',
-        type: { name: 'number' },
-        description: 'Menu offsetTop(px).',
-        defaultValue: undefined,
-        table: {
-            type: {
-                summary: 'number',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: 'undefined',
-            },
-        },
-        control: {
-            type: 'number',
-        },
-    },
-    width: {
-        name: 'width',
-        type: { name: 'number' },
-        description: 'Menu width(px).',
-        defaultValue: undefined,
-        table: {
-            type: {
-                summary: 'number',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: 'undefined',
-            },
-        },
-        control: {
-            type: 'number',
-        },
-    },
-    height: {
-        name: 'height',
-        type: { name: 'number' },
-        description: 'Menu height(px).',
-        defaultValue: undefined,
-        table: {
-            type: {
-                summary: 'number',
-            },
-            category: 'props',
-            defaultValue: {
-                summary: 'undefined',
-            },
-        },
-        control: {
-            type: 'number',
-        },
-    },
     alwaysShowMenu: {
         name: 'alwaysShowMenu',
         type: { name: 'boolean' },
